requirements.txt
thefuzz==0.22.1
PyYAML==6.0.2
requests==2.31.0